Set up the Side Panel
The side panel is functional, it displays every objects in a list
format. The objects can be Receipts, Products, Employees or Customers.
Data stuctures were created for each type of Object that can be used in
the program. If an object is selected from the list, it will display
some of the information on the pane (Currently only works for the Sale
Pane).