Add in displays to add in more object
Added in displays to add in more objects, each display will display at
the bottom. Currently only the Customer pane works but the storage
service isn't completed for the Pane. Finished the display of the Sale
Pane. Now it will display a reciept's contents. Currently the place
holders don't have all the information filled out, so it displays only
what is available.