Finish building the Front end.
The front end is complete. The panes for adding data and displaying data
is complete. All buttons work, from submiting data and retrieveing data.
The Add sales button allows products to be added to the Reciept by just
clicking on the Side bar. The product can be removed by clicking a
remove button that is created with the product info. All that is left to
do is connect the project to a MySQL database and work with the queries.