Work more with the database
Completed all instances where the application displays tuples from the
database. Data can be added to the database and new tuples are created.
The only instances that need to be completed are allowing users to add
and reduce how many products are in the inventory. Also allowing the
user to make sales still needs to be implemented.